Split admin controls
